Refactor the shortening code into a module package
I don't know if "module package" is the correct terminology, or whether
it should just be "package" or "sub-package".

Regardless, this change extracts the URL shortening code into a package
within the module so that the code, overall, is better organised, and
main.go isn't as loaded up with everything and anything, which shouldn't
be the case.

What's more, I'm trying to continue learning from @mattoddie by being
better organised and splitting out concerns where it makes sense to do
so.